use ruby-full. Bundle with 1.16.1. avoid auto reboot

+++ config-rp/scripts/install-coincooler.sh
@@ -17,28 +17,28 @@ sudo apt install -y nodejs libsqlite3-dev secure-delete xdotool
 # fix issue with openssl support a-la https://github.com/oleganza/btcruby/issues/29
 sudo ln -nfs /usr/lib/arm-linux-gnueabihf/libssl.so.1.0.2 /usr/lib/arm-linux-gnueabihf/libssl.so
 
-# install and configure bundler
-gem install bundler
-bundle config path .bundle
-bundle config bin ".bundle/binstubs"
-bundle config github.https true
-bundle config disable_shared_gems true
+# install and configure bundler at version 1.16.1
+sudo gem install bundler -v 1.16.1
+bundle _1.16.1_ config path .bundle
+bundle _1.16.1_ config bin ".bundle/binstubs"
+bundle _1.16.1_ config github.https true
+bundle _1.16.1_ config disable_shared_gems true
 
 # clone the coincooler repo and bundle
 git clone https://github.com/assafshomer/rpcc.git coincooler
 cd coincooler
-bundle
+bundle _1.16.1_
 
 # migrate sqlite3 db
-bundle exec rake db:migrate
-bundle exec rake db:migrate RAILS_ENV=test
-bundle exec rake db:migrate RAILS_ENV=production
+bundle _1.16.1_ exec rake db:migrate
+bundle _1.16.1_ exec rake db:migrate RAILS_ENV=test
+bundle _1.16.1_ exec rake db:migrate RAILS_ENV=production
 
 # precompile assets
-bundle exec rake assets:precompile
+bundle _1.16.1_ exec rake assets:precompile
 
 # run tests
-bundle exec rspec spec
+bundle _1.16.1_ exec rspec spec
 
 # copy over config files
 sudo cp ~/coincooler/config-rp/launchers/.aliases ~/.aliases
@@ -47,20 +47,7 @@ sudo cp ~/coincooler/config-rp/launchers/coincooler.desktop /usr/share/raspi-ui-
 sudo mv /usr/share/rpd-wallpaper/road.jpg /usr/share/rpd-wallpaper/road.jpg.old
 sudo cp ~/coincooler/app/assets/images/Bitcoin_in_ice_transparent_bgd.png /usr/share/rpd-wallpaper/road.jpg
 
-# quicklaunch hack
-read -d '' String <<"EOF"
-# coincooler
-source .aliases
-read -p "  Launch CoinCooler? (y/n) " ans
-if [[ $ans != "y" ]]; then
-  exit
-fi
-coincooler
-echo
-EOF
-
-# Save to ~/.bashrc
-echo -e "\n${String}" >> ~/.bashrc
+echo "source .aliases" >> ~/.bashrc
 
 #add purge script to crontab
 crontab -l > mycron
@@ -84,5 +71,7 @@ sudo systemctl start rng-tools
 ELAPSED_TIME=$(($SECONDS - $START_TIME))
 echo "\nFinished in $(($ELAPSED_TIME/60/60)) hr, $(($ELAPSED_TIME/60%60)) min, and $(($ELAPSED_TIME%60)) sec\n" >> ~/Documents/coincooler-installation.log
 
-# reboot
-sudo reboot
+read -p "  Reboot RaspberryPi (y/n) " ans
+if [[ $ans != "y" ]]; then
+  sudo reboot
+fi
